feat: add shared subagent execution admission
